requirements.txt
pandas==2.0.0
requests==2.28.2
boto3~=1.26.121
python-dotenv~=1.0.0
apache-airflow==2.5.2
apache-airflow-providers-amazon==7.3.0
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
zipp==3.15.0
